+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.3] - 2023-08-03
+
 ### Added
 
 - Pass HTML attributes to select choices
@@ -37,7 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Everything!
 
-[unreleased]: https://github.com/Ambiki/impulse_view_components/compare/v0.1.1...HEAD
+[unreleased]: https://github.com/Ambiki/impulse_view_components/compare/v0.1.3...HEAD
+[0.1.3]: https://github.com/Ambiki/impulse_view_components/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/Ambiki/impulse_view_components/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/Ambiki/impulse_view_components/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/Ambiki/impulse_view_components/releases/tag/v0.1.0